Small fixes
- do not parse amount input if fractional part ends with 0, otherwise
it's impossible to type in fractions that contain zeros. An incorrect
value may be saved in amount, eg if the following fraction 2.01 is
entered, and then the last 1 is deleted, the last parsed and saved
value will be 2.01 instead of 2.0.

- do not clone the value of sorting field in the state, but use a
reference to it. This line triggered an exception of unwrap on None
for some reason, and it's unclear what contained that None.

- reset selected accounts in the state as soon as possible in the
withdraw handler. This still may be racey.